optional-requirements.txt
python-magic
ruamel.yaml
SQLAlchemy<2
commonmark
lxml
asyncpg
aiosqlite
prometheus_client
setuptools
uvloop
python-olm
unpaddedbase64
pycryptodome
base58
